+++ dreamtools/__init__.py
@@ -32,6 +32,7 @@ from dream4.D4C1.scoring import D4C1
 from dream4.D4C2.scoring import D4C2
 from dream4.D4C3.scoring import D4C3
 
+from dream5.D5C1.scoring import D5C1
 from dream5.D5C2.scoring import D5C2
 
 from dream7.D7C1.scoring import D7C1